backend/pyproject.toml
[project]
name = "nexus-rag"
version = "1.0.0"
description = "NexusRAG — Enterprise Document Intelligence Platform with FastAPI + Next.js"
requires-python = ">=3.11"
license = {text = "MIT"}

dependencies = [
    # ── Core Framework ─────────────────────────────
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.9",
    "websockets>=12.0",

    # ── LangChain ──────────────────────────────────
    "langchain>=0.2.0",
    "langchain-core>=0.2.0",
    "langchain-community>=0.2.0",
    "langchain-google-genai>=1.0.0",
    "langchain-huggingface>=0.0.3",
    "langchain-text-splitters>=0.2.0",

    # ── Embeddings & Vector Store ──────────────────
    "sentence-transformers>=3.0.0",
    "faiss-cpu>=1.8.0",
    "rank-bm25>=0.2.2",

    # ── LLM Providers ─────────────────────────────
    "google-generativeai>=0.7.0",

    # ── Document Processing ────────────────────────
    "pypdf>=4.0.0",
    "pdfplumber>=0.11.0",
    "python-docx>=1.1.0",
    "openpyxl>=3.1.2",
    "xlrd>=2.0.1",
    "pandas>=2.2.0",
    "Pillow>=10.0.0",

    # ── Re-ranking ─────────────────────────────────
    "torch>=2.0.0",

    # ── Utilities ──────────────────────────────────
    "structlog>=24.1.0",
    "tenacity>=8.3.0",
    "numpy>=1.26.0",
    "aiofiles>=23.2.1",
    "httpx>=0.27.0",

    # ── Scientific Pipeline ────────────────────────────
    "opencv-python-headless>=4.9.0",
    "pymupdf>=1.24.0",
    "transformers>=4.40.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "black>=24.4.0",
]

[build-system]
requires = ["setuptools>=69.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = true